backend/requirements.txt
APScheduler==3.5.3
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-Cors==3.0.6
Flask-Migrate==2.3.0
slackclient==1.3.0
trueskill==0.4.5
psycopg2-binary==2.7.5
keras==2.2.4
tensorflow==1.11.0
numpy==1.15.2
pandas==0.23.4